move py and mapping to mapping dir
