+++ dump/mapping/bufr_sfcshp_prepbufr_mapping.yaml
@@ -10,6 +10,12 @@ bufr:
       query: "*/TYP"
 
     # MetaData
+    timestamp:
+      timeoffset:
+        timeOffset: "*/DHR"
+        transforms:
+          - scale: 3600
+        referenceTime: "2021-08-01T00:00:00Z"
     prepbufrDataLevelCategory:
       query: "*/CAT"
     obsTimeMinusCycleTime:
@@ -47,6 +53,10 @@ bufr:
       query: "*/T___INFO/T__EVENT{1}/TOB"
       transforms:
         - offset: 273.15
+    virtualTemperatureObsValue:
+      query: "*/T___INFO/T__EVENT{1}/TOB"
+      transforms:
+        - offset: 273.15
     specificHumidityObsValue:
       type: float
       query: "*/Q___INFO/Q__EVENT{1}/QOB"
@@ -68,6 +78,8 @@ bufr:
       query: "*/Z___INFO/Z__EVENT{1}/ZQM"`
     airTemperatureQualityMarker:
       query: "*/T___INFO/T__EVENT{1}/TQM"
+    virtualTemperatureQualityMarker:
+      query: "*/T___INFO/T__EVENT{1}/TQM"
     specificHumidityQualityMarker:
       query: "*/Q___INFO/Q__EVENT{1}/QQM"
     waterTemperatureQualityMarker:
@@ -83,6 +95,8 @@ bufr:
       transforms:
         - scale: 100
     airTemperatureObsError:
+      query: "*/T___INFO/T__BACKG/TOE"
+    virtualTemperatureObsError:
       query: "*/T___INFO/T__BACKG/TOE"
     relativeHumidityObsError:
       query: "*/Q___INFO/Q__BACKG/QOE"
@@ -162,7 +176,7 @@ encoder:
 
     # MetaData
     - name: "MetaData/dateTime"
-      source: variables/dateTime
+      source: variables/timestamp
       units: 'seconds since 1970-01-01T00:00:00Z'
       longName: 'dateTime'
 
@@ -181,11 +195,11 @@ encoder:
       source: variables/longitude
       longName: "Longitude"
       units: "degree_east"
-      range: [0, 360]
+      #range: [0, 360]
 
-    - name: "MetaData/sequenceNumber"
-      source: variables/sequenceNumber
-      longName: "Observation SubType"
+#    - name: "MetaData/sequenceNumber"
+#      source: variables/sequenceNumber
+#      longName: "Observation SubType"
 
     - name: "MetaData/stationIdentification"
       source: variables/stationIdentification
@@ -221,7 +235,7 @@ encoder:
       units: "m"
 
     - name: "ObsValue/airTemperature"
-      source: variables/sensibleTemperatureObsValue
+      source: variables/airTemperatureObsValue
       longName: "Air Temperature"
       units: "K"
 
@@ -255,7 +269,7 @@ encoder:
       longName: "Station Elevation Quality Marker"
 
     - name: "QualityMarker/airTemperature"
-      source: variables/sensibleTemperatureQualityMarker
+      source: variables/airTemperatureQualityMarker
       longName: "Air Temperature Quality Marker"
 
     - name: "QualityMarker/virtualTemperature"
@@ -281,7 +295,7 @@ encoder:
       units: "Pa"
 
     - name: "ObsError/airTemperature"
-      source: variables/sensibleTemperatureObsError
+      source: variables/airTemperatureObsError
       longName: "Temperature Error"
       units: "K"
