--- dump/mapping/iodatest_prepbufr_adpsfc_mapping.yaml
@@ -1,181 +1,0 @@
-# (C) Copyright 2023 NOAA/NWS/NCEP/EMC
-#
-# This software is licensed under the terms of the Apache Licence Version 2.0
-# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
-
-bufr:
-  variables:
-    # ObsType
-    observationType:
-      query: "*/TYP"
-
-    # MetaData
-    prepbufrDataLevelCategory:
-      query: "*/CAT"
-    obsTimeMinusCycleTime:
-      query: "*/DHR"
-    longitude:
-      query: "*/XOB"
-    latitude:
-      query: "*/YOB"
-    stationIdentification:
-      query: "*/SID"
-    pressure:
-      query: "*/P___INFO/P__EVENT{1}/POB"
-      transforms:
-        - scale: 100
-    height:
-      query: "*/Z___INFO/Z__EVENT{1}/ZOB"
-      type: float
-    stationElevation:
-      query: "*/ELV"
-      type: float
-
-    # ObsValue
-    stationPressureObsValue:
-      query: "*/P___INFO/P__EVENT{1}/POB"
-      transforms:
-        - scale: 100
-    heightObsValue:
-      query: "*/Z___INFO/Z__EVENT{1}/ZOB"
-      type: float
-    stationElevationObsValue:
-      query: "*/ELV"
-      type: float
-
-    # QualityMarker
-    stationPressureQualityMarker:
-      query: "*/P___INFO/P__EVENT{1}/PQM"
-    heightQualityMarker:
-      query: "*/Z___INFO/Z__EVENT{1}/ZQM"
-    stationElevationQualityMarker:
-      query: "*/Z___INFO/Z__EVENT{1}/ZQM"
-
-    # ObsError
-    stationPressureObsError:
-      query: "*/P___INFO/P__BACKG/POE"
-      transforms:
-        - scale: 100
-
-encoder:
-
-  globals:
-    - name: "data_format"
-      type: string
-      value: "prepbufr"
-
-    - name: "subsets"
-      type: string
-      value: "ADPSFC"        
-
-    - name: "source"
-      type: string
-      value: "prepBUFR"
-
-    - name: "data_type"
-      type: string
-      value: "ADPSFC"
-
-    - name: "data_description"
-      type: string
-      value: "ADPSFC_prepbufr"
-
-    - name: "data_provider"
-      type: string
-      value: "U.S. NOAA"
-
-  variables:
-
-    # ObsType
-    - name: "ObsType/stationPressure"
-      source: variables/observationType
-      longName: "Station Pressure ObsType"
-
-    - name: "ObsType/stationElevation"
-      source: variables/observationType
-      longName: "Station Elevation ObsType"
-
-    - name: "ObsType/height"
-      source: variables/observationType
-      longName: "Height ObsType"
-
-    # MetaData
-    - name: "MetaData/prepbufrDataLevelCategory"
-      source: variables/prepbufrDataLevelCategory
-      units: '1'
-      longName: "Prepbufr Data Level Category"
-
-    - name: "MetaData/dateTime"
-      source: variables/dateTime
-      units: 'seconds since 1970-01-01T00:00:00Z'
-      longName: 'dateTime'
-
-    - name: "MetaData/latitude"
-      source: variables/latitude
-      longName: "Latitude"
-      units: "degree_north"
-      range: [-90, 90]
-
-    - name: "MetaData/longitude"
-      source: variables/longitude
-      longName: "Longitude"
-      units: "degree_east"
-      range: [0, 360]
-
-    - name: "MetaData/sequenceNumber"
-      source: variables/sequenceNumber
-      longName: "Sequence Number (Obs Subtype)"
-
-    - name: "MetaData/stationIdentification"
-      source: variables/stationIdentification
-      longName: "Station Identification"
-
-    - name: "MetaData/pressure"
-      source: variables/pressure
-      longName: "Pressure"
-
-    - name: "MetaData/height"
-      source: variables/height
-      longName: "Height"
-      units: "m"
-
-    - name: "MetaData/stationElevation"
-      source: variables/stationElevation
-      longName: "Station Elevation"
-      units: "m"
-
-    # ObsValue
-    - name: "ObsValue/stationPressure"
-      source: variables/stationPressureObsValue
-      longName: "Station Pressure"
-      units: "Pa"
-
-    - name: "ObsValue/stationElevation"
-      source: variables/stationElevationObsValue
-      longName: "Station Pressure"
-      units: "m"
-
-    - name: "ObsValue/height"
-      source: variables/heightObsValue
-      longName: "height"
-      units: "m"
-
-    # QualityMarker
-    - name: "QualityMarker/stationPressure"
-      source: variables/stationPressureQualityMarker
-      longName: "Station Pressure Quality Marker"
-
-    - name: "QualityMarker/stationElevation"
-      source: variables/heightQualityMarker
-      longName: "StationElevation Quality Marker"
-
-    - name: "QualityMarker/height"
-      source: variables/heightQualityMarker
-      longName: "Height Quality Marker"
-
-    # ObsError
-    - name: "ObsError/stationPressure"
-      source: variables/stationPressureObsError
-      longName: "Station Pressure Error"
-      units: "Pa"
-
